Combines dim label queries (part 3)
